+++ Grbl_Esp32/src/UserOutput.h
@@ -40,7 +40,7 @@ namespace UserOutput {
 
     class AnalogOutput {
     public:
-        AnalogOutput(uint8_t number, Pin& pin, float pwm_frequency);
+        AnalogOutput(uint8_t number, Pin& pin, uint32_t pwm_frequency);
         bool     set_level(uint32_t numerator);
         uint32_t denominator() { return 1UL << _resolution_bits; };
 
@@ -51,7 +51,7 @@ namespace UserOutput {
         uint8_t  _number = UNDEFINED_OUTPUT;
         Pin&     _pin;
         uint8_t  _pwm_channel = -1;  // -1 means invalid or not setup
-        float    _pwm_frequency;
+        uint32_t _pwm_frequency;
         uint8_t  _resolution_bits;
         uint32_t _current_value;
     };
